requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
channels==3.0.4